+++ worldeditadditions_core/utils/param2.lua
@@ -1,0 +1,32 @@
+
+---
+-- @module worldeditadditions_core.param2
+
+-- //set <nodename>
+-- //set param2|p2 <param2>
+
+
+local function param2_to_dir(param2_type, param2)
+	if param2_type == "" then
+		return {
+			
+		}
+	else
+		return nil
+	end
+end
+
+local function dir_to_param2(dir)
+	
+end
+
+local function orient(param2, rotlist)
+	
+end
+
+
+
+
+return {
+	orient = orient
+}